+++ CHANGES.rst
@@ -5,7 +5,10 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add mosaic support in a separate profile. If mosaic is installed, this will be installed.
+  [cekk]
+- Fix tile visibility only when the product is installed
+  [cekk]
 
 
 1.0a1 (2017-03-31)